chore(DB): Populate DB with data
Populated the database with dummy data so features can be tested when needed. Also made a slight change to the database itself, removing the Messaging table and added a msgHistory attribute to the DPEnrole table. Added json files for audit log and logins. Also made a file called text.py to be used miscellaneous code and testing e.g. making a dummy ciphertext.